58: testing passed for multi-moduled code
