+++ oo-equivalence/equivalence-itests/pom.xml
@@ -14,6 +14,7 @@
     <modules>
         <module>test-cases</module>
         <module>test-module1</module>
+        <module>test-module2</module>
     </modules>
     
     <dependencies>
