--- oo-equivalence/equivalence-itests/test-cases/src/main/java/com/pragmaticobjects/oo/equivalence/itests/classes/GuestList.java
@@ -1,48 +1,0 @@
-/*-
- * ===========================================================================
- * equivalence-itests.test-cases
- * ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
- * Copyright (C) 2019 - 2023 Kapralov Sergey
- * ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- * 
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- * 
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- * ============================================================================
- */
-package com.pragmaticobjects.oo.equivalence.itests.classes;
-
-import java.util.HashSet;
-
-/**
- *
- * @author skapral
- */
-public class GuestList {
-    private final HashSet<String> guests;
-
-    public GuestList(HashSet<String> guests) {
-        this.guests = guests;
-    }
-
-    public final void invite(String guest) {
-        guests.add(guest);
-    }
-
-    public final boolean isInvited(String person) {
-        return guests.contains(person);
-    }
-}
