+++ oo-equivalence/equivalence-itests/test-module2/pom.xml
@@ -1,25 +1,22 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>com.pragmaticobjects.oo.equivalence.itests</groupId>
         <artifactId>equivalence-itests</artifactId>
         <version>0.0.0-SNAPSHOT</version>
     </parent>
-    <artifactId>test-cases</artifactId>
-    <packaging>jar</packaging>
+    <artifactId>test-module2</artifactId>
     <name>equivalence-itests.${project.artifactId}</name>
-
-    <properties>
-        <maven.compiler.release>20</maven.compiler.release>
-    </properties>
+    <packaging>jar</packaging>
 
     <dependencies>
         <dependency>
             <groupId>com.pragmaticobjects.oo.equivalence.itests</groupId>
-            <artifactId>test-module2</artifactId>
+            <artifactId>test-module1</artifactId>
             <version>${project.version}</version>
-            <scope>test</scope>
         </dependency>
     </dependencies>
 </project>
