+++ oo-equivalence/equivalence-itests/test-module2/src/main/java/com/pragmaticobjects/oo/equivalence/itests/classes/AnonymousClass.java
@@ -1,0 +1,62 @@
+/*-
+ * ===========================================================================
+ * equivalence-itests.test-module2
+ * ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+ * Copyright (C) 2019 - 2023 Kapralov Sergey
+ * ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ * ============================================================================
+ */
+package com.pragmaticobjects.oo.equivalence.itests.classes;
+
+import com.pragmaticobjects.oo.equivalence.base.EObjectHint;
+
+/**
+ *
+ * @author skapral
+ */
+public @EObjectHint(enabled = false) class AnonymousClass {
+    public interface Interface {
+        int value();
+    }
+    
+    public static Interface anonymousImplementation(int value) {
+        return new Interface() {
+            @Override
+            public final int value() {
+                return value;
+            }
+        };
+    }
+    
+    public static Interface anonymousMutableImplementation(int value) {
+        return new Interface() {
+            private int counter = 0;
+            
+            @Override
+            public final int value() {
+                counter++;
+                return value;
+            }
+        };
+    }
+    
+    public static final Class<?> anonymousType = anonymousImplementation(42).getClass();
+    public static final Class<?> anonymousMutableType = anonymousMutableImplementation(42).getClass();
+}
